Refactor: Reorganize db package structure
- Move migration logic from main package to db package
  - Rename runMigrations to db.Migrate for better clarity
  - Update embed path to work within db package
- Separate sqlc generated code into db/sqlc subpackage
  - Move db/*.go files to db/sqlc/
  - Update sqlc.yaml configuration
  - Update all import paths and references
- Update main.go and runn/runn.go to use new structure
- Remove duplicate migration code in runn package

This structure makes the migration logic reusable across packages
while keeping sqlc-generated code and migration logic organized.
