+++ main.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/stsysd/sougen/api"
 	"github.com/stsysd/sougen/config"
+	"github.com/stsysd/sougen/db"
 	"github.com/stsysd/sougen/store"
 )
 
@@ -14,7 +15,7 @@ func main() {
 	cfg := config.NewConfig()
 
 	// SQLiteストアの初期化（マイグレーション関数を渡す）
-	sqliteStore, err := store.NewSQLiteStore(cfg.DataDir, runMigrations)
+	sqliteStore, err := store.NewSQLiteStore(cfg.DataDir, db.Migrate)
 	if err != nil {
 		log.Fatalf("Failed to initialize SQLite store: %v", err)
 	}
